Added up command, reuse of view, and use existing view.
Added up to parent command bound to 'u'.

Navigating around created a lot of views so now the plugin defaults to
reusing the same view.  A new setting was added to control this.
Cmd+Enter/Ctrl+Enter can be used to force a new view.

If a dired view already exists for a directory, the dired command will
now set focus to it.